[Chore #160168203] Restructure application
