+++ src/App/App.test.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import { shallow, } from 'enzyme';
-import Login from '../../src/Login';
 
-describe('Login', () => {
+import App from './index';
+
+describe('App', () => {
   it('renders without crashing', () => {
-    shallow(<Login />);
+    shallow(<App />);
   });
 });
