+++ src/App/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from '../../Header';
-import Main from '../Main';
+import Header from '../Header/components';
+import Main from '../Main/components';
 
 const App = () => (
   <div>
