+++ src/Header/components/index.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link, } from 'react-router-dom';
+import './index.scss';
 
 // The Header creates links that can be used to navigate
 // between routes.
@@ -8,7 +9,7 @@ const Header = () => (
     <nav>
       <ul>
         <li><Link to="/">Home</Link></li>
-        <li><Link to="/login">Login</Link></li>
+        <li className="nav-right"><Link to="/login">Login</Link></li>
         <li><Link to="/signup">Signup</Link></li>
       </ul>
     </nav>
