+++ src/Home/components/index.scss
