+++ src/Main/components/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Switch, Route, } from 'react-router-dom';
-import Home from '../../Home';
-import Login from '../../Login';
+import Home from '../../Home/components';
+import Login from '../../Login/components';
 import Signup from '../../Signup/containers/Signup';
-import NotFound from '../Error_pages/page_not_found';
+import NotFound from '../../Error_pages/components/page_not_found';
 
 const Main = () => (
   <main>
