+++ src/Signup/components/Signup.scss
@@ -33,6 +33,8 @@ body {
     width: 65px;
     height: 60px;
     margin: 0 auto;
+    padding-bottom: 80px;
+    padding-right: 120px;
 }
 
 .logo img {
@@ -51,6 +53,7 @@ from {
 }
 
 .form input {
+    transition: 0.6s ease;
     outline: 0;
     border: 1px solid #CCC;
     box-sizing: border-box;
@@ -65,7 +68,8 @@ from {
 }
 
 button {
-    width: 60%;
+    transition: 0.6s ease;
+    width: 90%;
     background: #3498db;
     line-height: 40px;
     border: 0;
@@ -73,9 +77,10 @@ button {
     color: white;
     border-radius: 20px;
     font-weight: 600;
+    font-size: 13px;
     margin-top: 15px;
     margin-bottom: 50px;
-    text-emphasis: 600px
+    text-emphasis: 600px;
 }
 
 .card {
@@ -103,7 +108,7 @@ margin-right: 25px;
 }
 
 input:hover {
-    box-shadow: 0 2px 4px 0 #000000;
+    box-shadow: 0 2px 4px 0 #3498db;
 }
 
 button:hover {
