+++ src/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider, } from 'react-redux';
 import { BrowserRouter, } from 'react-router-dom';
-import App from './views/App';
-import store from './redux/store';
+import App from './App';
+import store from './store';
 
 
 if (process.env.NODE_ENV !== 'production') {
